requirement.txt
blessings==1.7
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
clip==1.0
dataclasses==0.8
filelock==3.4.1
ftfy==6.0.3
gpustat==0.6.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
isort==5.10.1
joblib==1.1.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nltk==3.6.7
numpy==1.19.2
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==21.3
Pillow==8.3.1
pip==21.2.2
protobuf==3.19.4
psutil==5.9.1
pyparsing==3.0.9
regex==2022.6.2
requests==2.27.1
rouge==1.0.0rc2
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.96
setuptools==58.0.4
six==1.16.0
tensorboardX==2.4.1
threadpoolctl==3.1.0
timm==0.5.4
tokenizers==0.10.3
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.64.0
transformers==4.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
wheel==0.37.1
zipp==3.6.0
